added dummy data for user skills and fixed bugs

+++ src/main/java/it/finki/tinki/service/SkillService.java
@@ -5,5 +5,5 @@ import it.finki.tinki.model.Skill;
 import java.util.List;
 
 public interface SkillService {
-    List<Skill> returnSkillsBasedOnId(List<Integer> skillIds);
+    List<Skill> returnSkillsBasedOnId(List<Long> skillIds);
 }
